add 5.1 profile and move prior prepared upgrade step for 502 to 503 into the new 503 to 510
