+++ plone/app/upgrade/__init__.py
@@ -23,6 +23,7 @@ class HiddenProducts(object):
             'plone.app.upgrade.v42',
             'plone.app.upgrade.v43',
             'plone.app.upgrade.v50',
+            'plone.app.upgrade.v51',
             ]
 
 # Make sure folks upgrading from Plone 2.1 see a helpful message telling them
