+++ plone/app/upgrade/v51/__init__.py
@@ -1,0 +1,1 @@
+# -*- coding: utf-8 -*-
