+++ plone/app/upgrade/v51/configure.zcml
@@ -1,0 +1,22 @@
+<configure
+    xmlns="http://namespaces.zope.org/zope"
+    xmlns:gs="http://namespaces.zope.org/genericsetup"
+    xmlns:zcml="http://namespaces.zope.org/zcml"
+    i18n_domain="plone">
+
+    <include file="profiles.zcml" />
+
+    <gs:upgradeSteps
+        source="5013"
+        destination="5101"
+        profile="Products.CMFPlone:plone">
+
+      <gs:upgradeStep
+           title="Run to51alpha1 upgrade profile."
+           description=""
+           handler=".alphas.to51alpha1"
+           />
+
+    </gs:upgradeSteps>
+
+</configure>
