+++ plone/app/upgrade/v51/profiles/to_alpha1/registry.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0"?>
+<registry>
+  <records
+      interface="Products.ResourceRegistries.interfaces.settings.IResourceRegistriesSettings"
+      prefix="plone">
+    <value
+        key="lessvariables"
+        purge="False">
+      <element key="plone-toolbar-submenu-width">180px</element>
+    </value>
+  </records>
+</registry>
